requirements.txt
Flask==3.0.0
cryptography==41.0.7
bcrypt==4.1.1
argon2-cffi==23.1.0
